requirements.txt
requests
pydantic
PySide6
python-dotenv